+++ src/components/add/NicknameModal/styles.ts
