+++ src/components/cardList/styles.ts
